+++ WHATS-NEW.md
@@ -15,6 +15,8 @@
 
 ## [Unreleased]
 
+## [0.14.6] - 2026-07-10
+
 ### New
 
 - **Wire a known service once and reuse it everywhere.** `flux endpoint add <name> --url <url>`